+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.16.0 ##
 
+### Features ###
+
+* Added numpad support to the ServerUtilities `TextBox` class.
+
 ### Bug Fixes ###
 
 * Fixed some issues with mixins that were causing injections to not properly get